// sourcecode/BTrak.Packages/snova-trading-module/trading-module-components/trading-components/src/lib/billing/models/programs-details.model.ts
export class ProgramDetails {
    programName: string;
    commencementDate: any;
    duration: any;
    image: any;
    country: any;
    phases: any;
    programId: any;
    dataSourceId: any;
    formData: any;
    formDataObject: any;
    contractType: any;
    userIds: any;
    description: any;
    programOwner: any;
    id : any;
    createdBy : any;
}